Вместо long - bigint

+++ src/main/resources/schema.sql
@@ -1,14 +1,14 @@
 
 CREATE TABLE IF NOT EXISTS electro_shop (
-shop_id INT NOT NULL,
-electro_item_id INT NOT NULL,
+shop_id BIGINT NOT NULL,
+electro_item_id BIGINT NOT NULL,
 count INT,
 CONSTRAINT electro_shop_pk PRIMARY KEY(shop_id, electro_item_id)
 );
 
 
 CREATE TABLE IF NOT EXISTS shop (
-id INT NOT NULL AUTO_INCREMENT,
+id BIGINT NOT NULL AUTO_INCREMENT,
 name VARCHAR(250) NOT NULL,
 address TEXT,
 CONSTRAINT shop_pk PRIMARY KEY(id)
@@ -16,10 +16,10 @@ CONSTRAINT shop_pk PRIMARY KEY(id)
 
 
 CREATE TABLE IF NOT EXISTS electro_item (
-id INT NOT NULL AUTO_INCREMENT,
+id BIGINT NOT NULL AUTO_INCREMENT,
 name VARCHAR(150) NOT NULL,
-e_type_id INT NOT NULL,
-price INT NOT NULL,
+e_type_id BIGINT NOT NULL,
+price BIGINT NOT NULL,
 count_ INT NOT NULL,  /*'возможные проблемы из-за имени'*/
 archive BOOL, /* 'BOOL and BOOLEAN are synonyms of TINYINT(1), Zero is false, anything else is true'*/
 description TEXT,
@@ -27,48 +27,48 @@ CONSTRAINT electro_item_pk PRIMARY KEY(id)
 );
 
 CREATE TABLE IF NOT EXISTS employee (
-id INT NOT NULL AUTO_INCREMENT,
+id BIGINT NOT NULL AUTO_INCREMENT,
 last_name VARCHAR(100) NOT NULL,
 first_name VARCHAR(100) NOT NULL,
 patronymyc VARCHAR(100) NOT NULL,
 birth_date DATE NOT NULL,
-position_id INT NOT NULL,
-shop_id INT NOT NULL,
+position_id BIGINT NOT NULL,
+shop_id BIGINT NOT NULL,
 gender BOOL,
 CONSTRAINT employee_pk PRIMARY KEY(id)
 );
 
 CREATE TABLE IF NOT EXISTS purchase (
 id INT NOT NULL AUTO_INCREMENT,
-electro_id INT NOT NULL,
-employee_id INT NOT NULL,
+electro_id BIGINT NOT NULL,
+employee_id BIGINT NOT NULL,
 purchase_date DATETIME NOT NULL,
-type_id INT NOT NULL,
-shop_id INT NOT NULL,
+type_id BIGINT NOT NULL,
+shop_id BIGINT NOT NULL,
 CONSTRAINT purchase_pk PRIMARY KEY(id)
 );
 
 
 CREATE TABLE IF NOT EXISTS electro_type (
-id INT NOT NULL AUTO_INCREMENT,
+id BIGINT NOT NULL AUTO_INCREMENT,
 name VARCHAR(150) NOT NULL,
 CONSTRAINT electro_type_pk PRIMARY KEY(id)
 );
 
 CREATE TABLE IF NOT EXISTS electro_employee (
-employee_id INT NOT NULL,
-electro_type_id INT NOT NULL,
+employee_id BIGINT NOT NULL,
+electro_type_id BIGINT NOT NULL,
 CONSTRAINT electro_employee_pk PRIMARY KEY(employee_id, electro_type_id)
 );
 
 CREATE TABLE IF NOT EXISTS position_type (
-id INT NOT NULL AUTO_INCREMENT,
+id BIGINT NOT NULL AUTO_INCREMENT,
 name VARCHAR(150) NOT NULL,
 CONSTRAINT position_type_pk PRIMARY KEY(id)
 );
 
 CREATE TABLE IF NOT EXISTS purchase_type (
-id INT NOT NULL AUTO_INCREMENT,
+id BIGINT NOT NULL AUTO_INCREMENT,
 name VARCHAR(150) NOT NULL,
 CONSTRAINT purchase_type_pk PRIMARY KEY(id)
 );
